+++ Media Blocks SDK/_CodeSnippets/udp-mpegts-streamer/App.xaml.cs
@@ -1,0 +1,12 @@
+using System.Windows;
+
+namespace udp_mpegts_streamer_mb
+{
+    /// <summary>
+    /// Interaction logic for App.xaml
+    /// </summary>
+    public partial class App : Application
+    {
+    }
+
+}